requirements.txt
Django==2.2.3
django-heroku==0.3.1
ONEmSDK==0.7.0
PyJWT==1.7.1
requests==2.22.0:w
